requirements.txt
python-dotenv==1.0.1
httpx==0.27.2
pandas==2.2.3
numpy==2.1.2
sqlalchemy==2.0.36
psycopg[binary]==3.2.3
pyarrow==17.0.0
orjson==3.10.7
typer==0.12.5
rich==13.9.2
yfinance==0.2.43
click==8.1.7